Criação da rotina de adição de pontos.

+++ DotzMVP.Lib/Infrastructure/Data/Model/Person.cs
@@ -9,7 +9,7 @@ namespace DotzMVP.Lib.Infrastructure.Data.Model
         public string Email { get; set; }
         public string Name { get; set; }
         public string Password { get; set; }
-        public int TotalScore { get; set; }
+        public double TotalScore { get; set; }
         public Guid? AddressID { get; set; }
         public Address Address { get; set; }
         public Guid? CustomerID { get; set; }
